// frontend/src/api/generated/gateways/gateways.ts
export type listGatewaysApiV1GatewaysGetResponse200 = {
  data: LimitOffsetPageTypeVarCustomizedGatewayRead;
  status: 200;
};

export type listGatewaysApiV1GatewaysGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type listGatewaysApiV1GatewaysGetResponseSuccess =
  listGatewaysApiV1GatewaysGetResponse200 & {
    headers: Headers;
  };
export type listGatewaysApiV1GatewaysGetResponseError =
  listGatewaysApiV1GatewaysGetResponse422 & {
    headers: Headers;
  };

export type listGatewaysApiV1GatewaysGetResponse =
  | listGatewaysApiV1GatewaysGetResponseSuccess
  | listGatewaysApiV1GatewaysGetResponseError;