// frontend/src/api/generated/model/gatewayRead.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * AxiaCraft Product Foundry API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Gateway payload returned from read endpoints.
 */
export interface GatewayRead {
  allow_insecure_tls?: boolean;
  created_at: string;
  disable_device_pairing?: boolean;
  id: string;
  name: string;
  organization_id: string;
  token?: string | null;
  updated_at: string;
  url: string;
  workspace_root: string;
}